Implementing WASD and Interact into level 1
