--- [SENDHELP] ARI/Assets/Developers/Jonah/Scripts/Interact/InteractLogic.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 99bb098602721ed4eab98f6c1ef67d27
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
