+++ [SENDHELP] ARI/Assets/Developers/Jonah/Scripts/Interact/Level1.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f3b3c2b9f5c6e04439334b06c0140587
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
