+++ [SENDHELP] ARI/Assets/Developers/Jonah/Scripts/Interact/Level1/Level1Interact.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 64bbed3272cdcfc45b8d548b05ebca95
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
