+++ [SENDHELP] ARI/Assets/Developers/Jonah/Scripts/Interact/Level1/Lvl1term1Interact.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f19dd5318566ea540a8346ee2057a7ef
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
