+++ [SENDHELP] ARI/Assets/Developers/Jonah/Scripts/Interact/Level1/Lvl1term2Interact.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 58765ad44306f9e45813484eb67362d4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
