+++ [SENDHELP] ARI/Assets/Developers/Jonah/Scripts/Interact/New Folder.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e942eb9a1eb3f604589751da26c8fbc9
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
