+++ [SENDHELP] ARI/Assets/Developers/Jonah/Scripts/Interact/Tutorial.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c4bc94f62a5788146a766910b92ec0c3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
