+++ [SENDHELP] ARI/Assets/Developers/Jonah/Scripts/disableColliders.cs
@@ -1,0 +1,19 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class disableColliders : MonoBehaviour
+{
+    public Collider sphereCollider;
+    // Start is called before the first frame update
+    void Start()
+    {
+        
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        sphereCollider.enabled = false;
+    }
+}
