+++ [SENDHELP] ARI/Assets/Developers/Jonah/Scripts/disableColliders.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2854a68bc137b324aa929453aad162f4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
