+++ [SENDHELP] ARI/Assets/Developers/Jonah/Scripts/interactables/Level1.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0af8c8e73043b1f4fa468c77392d2b74
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
