+++ [SENDHELP] ARI/Assets/Developers/Jonah/Scripts/interactables/Level1/Level1NotePads.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a95eaeb4afdaaa6469f1c94940a6b1de
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
