+++ [SENDHELP] ARI/Assets/Developers/Jonah/Scripts/interactables/Level1/Level1NotePads/LivingQuartersNote.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5bec25e2189e01f4692e29fb612d5b0b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
