+++ [SENDHELP] ARI/Assets/Developers/Jonah/Scripts/interactables/Level1/Level1NotePads/ServerNoteInteract.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b5424b9abd499a94885ad94db605b41d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
