+++ [SENDHELP] ARI/Assets/Developers/Jonah/Scripts/interactables/Level1/Level1NotePads/StorageRoomNote.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0752f93fdd244694e9d1a35c3aedc509
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
