+++ [SENDHELP] ARI/Assets/Developers/Jonah/Scripts/interactables/Level1/Level1Swipes.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cdfae78fa0e5015409ef6a415bb4f328
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
