+++ [SENDHELP] ARI/Assets/Developers/Jonah/Scripts/interactables/Level1/Level1Swipes/ServerRoomSwipe.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5ad5866855b80a743b4479ca855bb793
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
