+++ [SENDHELP] ARI/Assets/Developers/Jonah/Scripts/interactables/Level1/Level1Swipes/XHallSwipeInteract.cs
@@ -1,0 +1,18 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class XHallSwipeInteract : UpdatedInteractable
+{
+    // Start is called before the first frame update
+    void Start()
+    {
+        
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        
+    }
+}
