+++ [SENDHELP] ARI/Assets/Developers/Jonah/Scripts/interactables/Level1/Level1Swipes/XHallSwipeInteract.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d0bd7f24cb1ae804c96f143184052747
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
