+++ [SENDHELP] ARI/Assets/Developers/Jonah/Scripts/interactables/Level1/Level1Terminals.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bfe754c1a9d7b2c49a00f477fc8fded6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
