+++ [SENDHELP] ARI/Assets/Developers/Jonah/Scripts/interactables/Level1/Level1Terminals/GreenhouseTerminal.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f3016393f143c2d4ea925ce6138286cc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
