+++ [SENDHELP] ARI/Assets/Developers/Jonah/Scripts/interactables/Level1/Level1Terminals/LivingQuarterTerminal.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 00d36cb8074ba7446b0bdce809100308
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
