+++ [SENDHELP] ARI/Assets/Developers/Jonah/Scripts/interactables/Level1/Level1Terminals/ServerTerminalInteract.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4c52b26281a9ae14bab50ee092b5c195
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
