--- [SENDHELP] ARI/Assets/Developers/Jonah/Scripts/interactables/Level1Notepad.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cf65926d6c58df14d8569d6a72d6a127
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
