+++ [SENDHELP] ARI/Assets/Developers/Jonah/Scripts/interactables/Tutorial.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8d541de99102bec4fa2bc087ce6ddf65
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
