--- [SENDHELP] ARI/Assets/Developers/Jonah/Scripts/interactables/TutorialChipNode.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6a5e491766661c542b4a7e1f78e4a7ca
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
