--- [SENDHELP] ARI/Assets/Developers/Jonah/Scripts/interactables/TutorialTerminal.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d05f74cbdf566cf468a23370261390eb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
